+++ internal/sdkprovider/kafkatopicrepository/update.go
@@ -3,9 +3,9 @@ package kafkatopicrepository
 import (
 	"context"
 
-	"github.com/aiven/aiven-go-client/v2"
+	"github.com/aiven/go-client-codegen/handler/kafkatopic"
 )
 
-func (rep *repository) Update(ctx context.Context, project, service, topic string, req aiven.UpdateKafkaTopicRequest) error {
-	return rep.client.Update(ctx, project, service, topic, req)
+func (rep *repository) Update(ctx context.Context, project, service, topic string, req *kafkatopic.ServiceKafkaTopicUpdateIn) error {
+	return rep.client.ServiceKafkaTopicUpdate(ctx, project, service, topic, req)
 }