+++ CHANGELOG.md
@@ -12,6 +12,7 @@ nav_order: 1
 
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 
+- Migrate `aiven_kafka_topic` resource and data source to the generated Go client
 - Add `aiven_organization_application_user` field `create_time`: Time this application user was created.
 - Add `aiven_kafka_topic` field `config.message_timestamp_after_max_ms`: The maximum difference allowed between
   the timestamp when a broker receives a message and the timestamp specified in the message. Applies only for messages